# openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CoverageResponse": {
        "properties": {
          "end_date": {
            "title": "End Date",
            "type": "string"
          },
          "field_coverage": {
            "items": {
              "$ref": "#/components/schemas/FieldCoverage"
            },
            "title": "Field Coverage",
            "type": "array"
          },
          "n_days": {
            "title": "N Days",
            "type": "integer"
          },
          "n_tickers": {
            "title": "N Tickers",
            "type": "integer"
          },
          "ohlcv_coverage_pct": {
            "title": "Ohlcv Coverage Pct",
            "type": "number"
          },
          "ohlcv_missing_cells": {
            "title": "Ohlcv Missing Cells",
            "type": "integer"
          },
          "ohlcv_total_cells": {
            "title": "Ohlcv Total Cells",
            "type": "integer"
          },
          "start_date": {
            "title": "Start Date",
            "type": "string"
          },
          "ticker_coverage": {
            "items": {
              "$ref": "#/components/schemas/TickerCoverage"
            },
            "title": "Ticker Coverage",
            "type": "array"
          },
          "universe_id": {
            "title": "Universe Id",
            "type": "string"
          }
        },
        "required": [
          "universe_id",
          "n_tickers",
          "n_days",
          "start_date",
          "end_date",
          "ohlcv_total_cells",
          "ohlcv_missing_cells",
          "ohlcv_coverage_pct",
          "field_coverage",
          "ticker_coverage"
        ],
        "title": "CoverageResponse",
        "type": "object"
      },
...
      "FieldCoverage": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "fill_rate": {
            "title": "Fill Rate",
            "type": "number"
          },
          "n_present": {
            "title": "N Present",
            "type": "integer"
          },
          "n_total": {
            "title": "N Total",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "name",
          "category",
          "tier",
          "fill_rate",
          "n_present",
          "n_total"
        ],
        "title": "FieldCoverage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperandCatalogResponse": {
        "description": "Augmented WorldQuant catalog. Each item carries `tier` and `implemented`.\n\nTier semantics:\n  - \"T1\" \u2014 implemented now, included in factor expressions today\n  - \"T2\" \u2014 implemented, requires v2 panel data (sector / fundamentals)\n  - \"T3\" \u2014 catalog-only; data source unavailable or operator semantics\n    out of scope (vector ops, news/options/sentiment, infix comparisons)",
        "properties": {
          "operands": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Operands",
            "type": "array"
          },
          "operators": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Operators",
            "type": "array"
          },
          "tier_summary": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "title": "Tier Summary",
            "type": "object"
          }
        },
        "required": [
          "operators",
          "operands",
          "tier_summary"
        ],
        "title": "OperandCatalogResponse",
        "type": "object"
      },
...
      "TickerCoverage": {
        "properties": {
          "fill_rate": {
            "title": "Fill Rate",
            "type": "number"
          },
          "n_missing": {
            "title": "N Missing",
            "type": "integer"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "fill_rate",
          "n_missing"
        ],
        "title": "TickerCoverage",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AlphaCore Dashboard API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/data/coverage": {
      "get": {
        "description": "Return per-field fill-rate stats for the active panel.\n\nOriginal v1 returned a (T, N) matrix for a heatmap, which was visually\npretty but information-dense at zero \u2014 every cell green for cleaned\nOHLCV. v2 returns (a) per-operand fill rates so fundamentals' real\n73-98% variation is visible and (b) per-ticker OHLCV coverage as\nhorizontal bars instead of a wall of cells.",
        "operationId": "get_coverage_api_v1_data_coverage_get",
        "parameters": [
          {
            "in": "query",
            "name": "universe_id",
            "required": false,
            "schema": {
              "const": "SP500_subset",
              "default": "SP500_subset",
              "title": "Universe Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Coverage",
        "tags": [
          "data"
        ]
      }
    },
    "/api/v1/data/operands": {
      "get": {
        "description": "Return the augmented catalog of operators and operands.\n\n`tier` filter:\n  - \"all\":       everything (default \u2014 UI catalog page)\n  - \"available\": only items with implemented=true (LLM prompt feed)\n  - \"T1\"/\"T2\"/\"T3\": exact tier match\n\n`include_signature=true` enriches each operator entry with `signature` and\n`source_snippet` fields pulled from scan/vectorized.py via inspect. Used\nby the Methodology page (B1) to render the actual NumPy implementation\nbody rather than just the doc string.",
        "operationId": "list_operands_api_v1_data_operands_get",
        "parameters": [
          {
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "available",
                "T1",
                "T2",
                "T3"
              ],
              "title": "Tier",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_signature",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Signature",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperandCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Operands",
        "tags": [
          "data"
        ]
      }
    },
...
  }
}